refactor: refactored commit with new bag, after first running benchmark I can't start second time and others

+++ app/src/main/java/com/example/benchmarks/ui/benchmark/BenchmarkFragment.java
@@ -72,7 +72,7 @@ public class BenchmarkFragment extends Fragment implements View.OnClickListener,
 
     @Override
     public void onFragmentResult(@NonNull String requestKey, @NonNull Bundle result) {
-        viewModel.setSizeCollectionLiveData(result.getLong(InputFragment.LONG_COLLECTION_SIZE_KEY));
+        viewModel.setSizeCollectionLiveData(result.getInt(InputFragment.LONG_COLLECTION_SIZE_KEY));
         String size = result.getString(InputFragment.STRING_COLLECTION_SIZE_KEY);
         bind.edCollectionsFragment.setText(size);
     }
